requirements.txt
oracledb
pymongo
python-dotenv